+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3 - themeMode, theme, and darkTheme
+
+- Exposing themeMode, theme, and darkTheme through custom widget.
+
+## 0.0.2 - Widgetbook support
+
+- Added widgetbook support in dart dev tools.
+
 ## 0.0.1 - Initial Release
 
 - Displays the `primaryColor` from your `MaterialApp` theme inside Dart DevTools.